Fix up tests, begin new API
