+++ antlr/build.gradle.kts
@@ -1,0 +1,4 @@
+dependencies {
+    // https://mvnrepository.com/artifact/org.antlr/antlr4-runtime
+    implementation("org.antlr:antlr4-runtime:4.8-1")
+}
