+++ antlr/src/main/java/org/jglrxavpok/hephaistos/antlr/SNBT.interp
@@ -1,0 +1,71 @@
+token literal names:
+null
+'{'
+','
+'}'
+':'
+'['
+']'
+'B'
+';'
+'I'
+'L'
+null
+null
+null
+'-'
+null
+null
+null
+null
+null
+null
+null
+null
+
+token symbolic names:
+null
+null
+null
+null
+null
+null
+null
+null
+null
+null
+null
+DoubleQuoteText
+SingleQuoteText
+BOOLEAN
+NEGATIVE_SIGN
+FLOAT
+DOUBLE
+INTEGER
+LONG
+BYTE
+SHORT
+WS
+IDENTIFIER_LETTERS
+
+rule names:
+snbt
+element
+compound
+namedElement
+list
+byteArray
+intArray
+longArray
+doubleNBT
+floatNBT
+longNBT
+byteNBT
+shortNBT
+intNBT
+stringNBT
+identifier
+
+
+atn:
+[3, 24715, 42794, 33075, 47597, 16764, 15335, 30598, 22884, 3, 24, 164, 4, 2, 9, 2, 4, 3, 9, 3, 4, 4, 9, 4, 4, 5, 9, 5, 4, 6, 9, 6, 4, 7, 9, 7, 4, 8, 9, 8, 4, 9, 9, 9, 4, 10, 9, 10, 4, 11, 9, 11, 4, 12, 9, 12, 4, 13, 9, 13, 4, 14, 9, 14, 4, 15, 9, 15, 4, 16, 9, 16, 4, 17, 9, 17, 3, 2, 3, 2, 3, 2, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 5, 3, 50, 10, 3, 3, 4, 3, 4, 3, 4, 3, 4, 7, 4, 56, 10, 4, 12, 4, 14, 4, 59, 11, 4, 5, 4, 61, 10, 4, 3, 4, 5, 4, 64, 10, 4, 3, 4, 3, 4, 3, 5, 3, 5, 3, 5, 3, 5, 3, 6, 3, 6, 3, 6, 3, 6, 7, 6, 76, 10, 6, 12, 6, 14, 6, 79, 11, 6, 5, 6, 81, 10, 6, 3, 6, 5, 6, 84, 10, 6, 3, 6, 3, 6, 3, 7, 3, 7, 3, 7, 3, 7, 3, 7, 3, 7, 7, 7, 94, 10, 7, 12, 7, 14, 7, 97, 11, 7, 3, 7, 5, 7, 100, 10, 7, 5, 7, 102, 10, 7, 3, 7, 3, 7, 3, 8, 3, 8, 3, 8, 3, 8, 3, 8, 3, 8, 7, 8, 112, 10, 8, 12, 8, 14, 8, 115, 11, 8, 3, 8, 5, 8, 118, 10, 8, 5, 8, 120, 10, 8, 3, 8, 3, 8, 3, 9, 3, 9, 3, 9, 3, 9, 3, 9, 3, 9, 7, 9, 130, 10, 9, 12, 9, 14, 9, 133, 11, 9, 3, 9, 5, 9, 136, 10, 9, 5, 9, 138, 10, 9, 3, 9, 3, 9, 3, 10, 3, 10, 3, 11, 3, 11, 3, 12, 3, 12, 3, 13, 3, 13, 3, 14, 3, 14, 3, 15, 3, 15, 3, 16, 3, 16, 3, 16, 5, 16, 157, 10, 16, 3, 17, 6, 17, 160, 10, 17, 13, 17, 14, 17, 161, 3, 17, 2, 2, 18, 2, 4, 6, 8, 10, 12, 14, 16, 18, 20, 22, 24, 26, 28, 30, 32, 2, 3, 4, 2, 15, 15, 21, 21, 2, 176, 2, 34, 3, 2, 2, 2, 4, 49, 3, 2, 2, 2, 6, 51, 3, 2, 2, 2, 8, 67, 3, 2, 2, 2, 10, 71, 3, 2, 2, 2, 12, 87, 3, 2, 2, 2, 14, 105, 3, 2, 2, 2, 16, 123, 3, 2, 2, 2, 18, 141, 3, 2, 2, 2, 20, 143, 3, 2, 2, 2, 22, 145, 3, 2, 2, 2, 24, 147, 3, 2, 2, 2, 26, 149, 3, 2, 2, 2, 28, 151, 3, 2, 2, 2, 30, 156, 3, 2, 2, 2, 32, 159, 3, 2, 2, 2, 34, 35, 5, 4, 3, 2, 35, 36, 7, 2, 2, 3, 36, 3, 3, 2, 2, 2, 37, 50, 5, 24, 13, 2, 38, 50, 5, 20, 11, 2, 39, 50, 5, 26, 14, 2, 40, 50, 5, 22, 12, 2, 41, 50, 5, 28, 15, 2, 42, 50, 5, 18, 10, 2, 43, 50, 5, 30, 16, 2, 44, 50, 5, 12, 7, 2, 45, 50, 5, 14, 8, 2, 46, 50, 5, 16, 9, 2, 47, 50, 5, 10, 6, 2, 48, 50, 5, 6, 4, 2, 49, 37, 3, 2, 2, 2, 49, 38, 3, 2, 2, 2, 49, 39, 3, 2, 2, 2, 49, 40, 3, 2, 2, 2, 49, 41, 3, 2, 2, 2, 49, 42, 3, 2, 2, 2, 49, 43, 3, 2, 2, 2, 49, 44, 3, 2, 2, 2, 49, 45, 3, 2, 2, 2, 49, 46, 3, 2, 2, 2, 49, 47, 3, 2, 2, 2, 49, 48, 3, 2, 2, 2, 50, 5, 3, 2, 2, 2, 51, 60, 7, 3, 2, 2, 52, 57, 5, 8, 5, 2, 53, 54, 7, 4, 2, 2, 54, 56, 5, 8, 5, 2, 55, 53, 3, 2, 2, 2, 56, 59, 3, 2, 2, 2, 57, 55, 3, 2, 2, 2, 57, 58, 3, 2, 2, 2, 58, 61, 3, 2, 2, 2, 59, 57, 3, 2, 2, 2, 60, 52, 3, 2, 2, 2, 60, 61, 3, 2, 2, 2, 61, 63, 3, 2, 2, 2, 62, 64, 7, 4, 2, 2, 63, 62, 3, 2, 2, 2, 63, 64, 3, 2, 2, 2, 64, 65, 3, 2, 2, 2, 65, 66, 7, 5, 2, 2, 66, 7, 3, 2, 2, 2, 67, 68, 5, 30, 16, 2, 68, 69, 7, 6, 2, 2, 69, 70, 5, 4, 3, 2, 70, 9, 3, 2, 2, 2, 71, 80, 7, 7, 2, 2, 72, 77, 5, 4, 3, 2, 73, 74, 7, 4, 2, 2, 74, 76, 5, 4, 3, 2, 75, 73, 3, 2, 2, 2, 76, 79, 3, 2, 2, 2, 77, 75, 3, 2, 2, 2, 77, 78, 3, 2, 2, 2, 78, 81, 3, 2, 2, 2, 79, 77, 3, 2, 2, 2, 80, 72, 3, 2, 2, 2, 80, 81, 3, 2, 2, 2, 81, 83, 3, 2, 2, 2, 82, 84, 7, 4, 2, 2, 83, 82, 3, 2, 2, 2, 83, 84, 3, 2, 2, 2, 84, 85, 3, 2, 2, 2, 85, 86, 7, 8, 2, 2, 86, 11, 3, 2, 2, 2, 87, 88, 7, 7, 2, 2, 88, 89, 7, 9, 2, 2, 89, 101, 7, 10, 2, 2, 90, 95, 5, 24, 13, 2, 91, 92, 7, 4, 2, 2, 92, 94, 5, 24, 13, 2, 93, 91, 3, 2, 2, 2, 94, 97, 3, 2, 2, 2, 95, 93, 3, 2, 2, 2, 95, 96, 3, 2, 2, 2, 96, 99, 3, 2, 2, 2, 97, 95, 3, 2, 2, 2, 98, 100, 7, 4, 2, 2, 99, 98, 3, 2, 2, 2, 99, 100, 3, 2, 2, 2, 100, 102, 3, 2, 2, 2, 101, 90, 3, 2, 2, 2, 101, 102, 3, 2, 2, 2, 102, 103, 3, 2, 2, 2, 103, 104, 7, 8, 2, 2, 104, 13, 3, 2, 2, 2, 105, 106, 7, 7, 2, 2, 106, 107, 7, 11, 2, 2, 107, 119, 7, 10, 2, 2, 108, 113, 5, 28, 15, 2, 109, 110, 7, 4, 2, 2, 110, 112, 5, 28, 15, 2, 111, 109, 3, 2, 2, 2, 112, 115, 3, 2, 2, 2, 113, 111, 3, 2, 2, 2, 113, 114, 3, 2, 2, 2, 114, 117, 3, 2, 2, 2, 115, 113, 3, 2, 2, 2, 116, 118, 7, 4, 2, 2, 117, 116, 3, 2, 2, 2, 117, 118, 3, 2, 2, 2, 118, 120, 3, 2, 2, 2, 119, 108, 3, 2, 2, 2, 119, 120, 3, 2, 2, 2, 120, 121, 3, 2, 2, 2, 121, 122, 7, 8, 2, 2, 122, 15, 3, 2, 2, 2, 123, 124, 7, 7, 2, 2, 124, 125, 7, 12, 2, 2, 125, 137, 7, 10, 2, 2, 126, 131, 5, 22, 12, 2, 127, 128, 7, 4, 2, 2, 128, 130, 5, 22, 12, 2, 129, 127, 3, 2, 2, 2, 130, 133, 3, 2, 2, 2, 131, 129, 3, 2, 2, 2, 131, 132, 3, 2, 2, 2, 132, 135, 3, 2, 2, 2, 133, 131, 3, 2, 2, 2, 134, 136, 7, 4, 2, 2, 135, 134, 3, 2, 2, 2, 135, 136, 3, 2, 2, 2, 136, 138, 3, 2, 2, 2, 137, 126, 3, 2, 2, 2, 137, 138, 3, 2, 2, 2, 138, 139, 3, 2, 2, 2, 139, 140, 7, 8, 2, 2, 140, 17, 3, 2, 2, 2, 141, 142, 7, 18, 2, 2, 142, 19, 3, 2, 2, 2, 143, 144, 7, 17, 2, 2, 144, 21, 3, 2, 2, 2, 145, 146, 7, 20, 2, 2, 146, 23, 3, 2, 2, 2, 147, 148, 9, 2, 2, 2, 148, 25, 3, 2, 2, 2, 149, 150, 7, 22, 2, 2, 150, 27, 3, 2, 2, 2, 151, 152, 7, 19, 2, 2, 152, 29, 3, 2, 2, 2, 153, 157, 5, 32, 17, 2, 154, 157, 7, 13, 2, 2, 155, 157, 7, 14, 2, 2, 156, 153, 3, 2, 2, 2, 156, 154, 3, 2, 2, 2, 156, 155, 3, 2, 2, 2, 157, 31, 3, 2, 2, 2, 158, 160, 7, 24, 2, 2, 159, 158, 3, 2, 2, 2, 160, 161, 3, 2, 2, 2, 161, 159, 3, 2, 2, 2, 161, 162, 3, 2, 2, 2, 162, 33, 3, 2, 2, 2, 20, 49, 57, 60, 63, 77, 80, 83, 95, 99, 101, 113, 117, 119, 131, 135, 137, 156, 161]
