--- common/src/antlr/gen/org/jglrxavpok/hephaistos/antlr/SNBT.tokens
@@ -1,33 +1,0 @@
-T__0=1
-T__1=2
-T__2=3
-T__3=4
-T__4=5
-T__5=6
-T__6=7
-T__7=8
-T__8=9
-T__9=10
-DoubleQuoteText=11
-SingleQuoteText=12
-BOOLEAN=13
-NEGATIVE_SIGN=14
-FLOAT=15
-DOUBLE=16
-INTEGER=17
-LONG=18
-BYTE=19
-SHORT=20
-WS=21
-IDENTIFIER_LETTERS=22
-'{'=1
-','=2
-'}'=3
-':'=4
-'['=5
-']'=6
-'B'=7
-';'=8
-'I'=9
-'L'=10
-'-'=14
