+++ common/src/main/kotlin/org/jglrxavpok/hephaistos/nbt/NBTNumber.kt
@@ -20,6 +20,7 @@ abstract class NBTNumber<Type: Number>(open val value: Type): NBT {
         return true
     }
 
+    @Deprecated("NBTNumbers are immutable", ReplaceWith("this"))
     final override fun deepClone() = this
 
 }
