+++ common/src/test/java/mca/MCALoading.java
@@ -3,9 +3,7 @@ package mca;
 import org.jglrxavpok.hephaistos.mca.AnvilException;
 import org.jglrxavpok.hephaistos.mca.ChunkColumn;
 import org.jglrxavpok.hephaistos.mca.RegionFile;
-import org.junit.jupiter.api.AfterAll;
-import org.junit.jupiter.api.BeforeAll;
-import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.*;
 
 import java.io.File;
 import java.io.IOException;
@@ -21,7 +19,7 @@ public class MCALoading {
 
     private RandomAccessFile file;
 
-    @BeforeAll
+    @BeforeEach
     public void init() throws IOException {
         Files.copy(Paths.get("src/test/resources/r.0.0.mca"), Paths.get("tmp_r.0.0.mca"), REPLACE_EXISTING);
         file = new RandomAccessFile(new File("tmp_r.0.0.mca"), "rw");
@@ -43,7 +41,7 @@ public class MCALoading {
         }
     }
 
-    @AfterAll
+    @AfterEach
     public void clean() throws IOException {
         try {
             file.close();
