+++ common/src/test/java/mca/MCASaving.java
@@ -4,15 +4,13 @@ import org.jglrxavpok.hephaistos.mca.AnvilException;
 import org.jglrxavpok.hephaistos.mca.BlockState;
 import org.jglrxavpok.hephaistos.mca.ChunkColumn;
 import org.jglrxavpok.hephaistos.mca.RegionFile;
-import org.junit.jupiter.api.AfterAll;
-import org.junit.jupiter.api.BeforeAll;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.*;
 
 import java.io.File;
 import java.io.IOException;
 import java.io.RandomAccessFile;
 import java.nio.file.Files;
+import java.nio.file.Path;
 import java.nio.file.Paths;
 
 import static java.nio.file.StandardCopyOption.REPLACE_EXISTING;
@@ -246,7 +244,7 @@ public class MCASaving {
         assertEquals((byte) 12, c.getSections()[0].getSkyLight(0,5,0));
     }
 
-    @AfterAll
+    @AfterEach
     public void clean() throws IOException {
         try {
             file.close();
