+++ gson/src/test/java/TestWriterPrimitives.java
@@ -1,23 +1,23 @@
 import com.google.gson.JsonElement;
 import org.jglrxavpok.hephaistos.json.NBTGsonWriter;
 import org.jglrxavpok.hephaistos.nbt.*;
-import org.junit.Before;
-import org.junit.Test;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
 
-import static org.junit.Assert.*;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
 public class TestWriterPrimitives {
 
     private NBTGsonWriter writer;
 
-    @Before
+    @BeforeEach
     public void init() {
         writer = new NBTGsonWriter();
     }
 
     @Test
     public void writeByte() {
-        JsonElement element = writer.write(new NBTByte((byte) 42));
+        JsonElement element = writer.write(NBT.Byte((byte) 42));
         assertEquals("42", element.toString());
     }
