+++ settings.gradle.kts
@@ -3,3 +3,4 @@ rootProject.name = "Hephaistos"
 include(":example")
 include(":gson")
 include(":common")
+include(":antlr")
